MeteoMeter/Ionic/src/pages/account/main/account-main.spec.ts

import { IonicModule, NavController, Platform } from 'ionic-angular';
import { NavMock } from '../../../../test-config/mocks/navMock';
import { PlatformMock } from '../../../../test-config/mocks/platformMock';
import { SplashScreen } from '@ionic-native/splash-screen';
import { SplashScreenMock } from '../../../../test-config/mocks/splashScreenMock';
import { StatusBar } from '@ionic-native/status-bar';
import { StatusBarMock } from '../../../../test-config/mocks/statusBarMock';
import { Title } from '@angular/platform-browser';
import { TitleMock } from '../../../../test-config/mocks/titleMock';
import { async, ComponentFixture, fakeAsync, flushMicrotasks, TestBed } from '@angular/core/testing';
import { TranslateModule, TranslateService } from '@ngx-translate/core';
import { TranslateServiceMock } from '../../../../test-config/mocks/translateServiceMock';
import { LanguagePageModule } from '../../language/language.module';
import { LanguagePage } from '../../language/language';
import { MeteoMeter } from '../../../app/app.component';
import { AccountMain } from "./account-main";
import { FullUseraaaaaaaaaaaaaa } from '../../../viewModels/fullUseraaaaaaaaaaaaaa';
import { UserAAAAA } from '../../../services/userAAAAA.service';
import { UserAAAAAMock } from '../../../../test-config/mocks/userAAAAAMock';
describe('AccountMain Component', () => {
  let fixture: ComponentFixture<AccountMain>;
  let component: AccountMain;
  let spy: jasmine.Spy;

  beforeEach(async(() => {
    TestBed.configureTestingModule({
      declarations: [
        MeteoMeter,
        AccountMain
      ],
      imports: [
        LanguagePageModule,
        TranslateModule.forChild(),
        IonicModule.forRoot(MeteoMeter)
      ],
      providers: [
        { provide: UserAAAAA, useClass: UserAAAAAMock },
        { provide: Platform, useClass: PlatformMock },
        { provide: StatusBar, useClass: StatusBarMock },
        { provide: SplashScreen, useClass: SplashScreenMock },
        { provide: TranslateService, useClass: TranslateServiceMock },
        { provide: NavController, useClass: NavMock },
        { provide: Title, useClass: TitleMock },
      ]
    }).compileComponents();
  }));

  beforeEach(async() => {
    fixture = TestBed.createComponent(AccountMain);
    component = fixture.componentInstance;
  });
  
  afterEach(() => {
    fixture.destroy();
    component = null;
  });

  it('AccountMain Component: should be created', fakeAsync(() => {
    expect(component instanceof AccountMain).toBeTruthy();
  }));

  it('AccountMain Component: should navigate to language page', fakeAsync(() => {
    spy = spyOn(component, 'onLanguage').and.callThrough();
    component.onLanguage();
    flushMicrotasks();
    expect(spy).toHaveBeenCalled();
    spy = spyOn(component._nav, 'push').and.callThrough();
    component.onLanguage();
    flushMicrotasks();
    expect(spy).toHaveBeenCalledWith(LanguagePage);
    spy.calls.mostRecent().returnValue.then((result) => {
      expect(result).toEqual('PUSH');
    });
  }));
  it('AccountMain Component - onSignOut: should do nothing', fakeAsync(() => {
    spy = spyOn(component, 'onSignOut').and.callThrough();
    component.onSignOut();
    flushMicrotasks();
    expect(spy).toHaveBeenCalled();
    var result = spy.calls.mostRecent().returnValue;
    expect(result).toBeUndefined();
  }));

  it('AccountMain Component - onGetUser: should use userAAAAA service to GetUser', fakeAsync(() => {
    var userProfileId = 'test';
    spy = spyOn(component, 'onGetUser').and.callThrough();
    component.onGetUser(
    userProfileId
    );
    flushMicrotasks();
    expect(spy).toHaveBeenCalledWith(
    userProfileId
    );
    spy = spyOn(component._userAAAAA, 'getFullInfos').and.callThrough();
    component.onGetUser(
    userProfileId
    );
    flushMicrotasks();
    expect(spy).toHaveBeenCalledWith(
    userProfileId
    );
    var result = spy.calls.mostRecent().returnValue;
    expect(result).toBeDefined();
  }));
  it('AccountMain Component - onToEditProfile: should navigate to account-editProfile page', fakeAsync(() => {
    spy = spyOn(component, 'onToEditProfile').and.callThrough();
    component.onToEditProfile();
    flushMicrotasks();
    expect(spy).toHaveBeenCalled();
    spy = spyOn(component._nav, 'push').and.callThrough();
    component.onToEditProfile();
    flushMicrotasks();
    expect(spy).toHaveBeenCalledWith('account-editProfile');
    spy.calls.mostRecent().returnValue.then((result) => {
      expect(result).toEqual('PUSH');
    });
  }));

  it('AccountMain Component - onToManageGroups: should navigate to manageGroups-main page', fakeAsync(() => {
    spy = spyOn(component, 'onToManageGroups').and.callThrough();
    component.onToManageGroups();
    flushMicrotasks();
    expect(spy).toHaveBeenCalled();
    spy = spyOn(component._nav, 'push').and.callThrough();
    component.onToManageGroups();
    flushMicrotasks();
    expect(spy).toHaveBeenCalledWith('manageGroups-main');
    spy.calls.mostRecent().returnValue.then((result) => {
      expect(result).toEqual('PUSH');
    });
  }));

  it('AccountMain Component - onToSettings: should navigate to settings-main page', fakeAsync(() => {
    spy = spyOn(component, 'onToSettings').and.callThrough();
    component.onToSettings();
    flushMicrotasks();
    expect(spy).toHaveBeenCalled();
    spy = spyOn(component._nav, 'push').and.callThrough();
    component.onToSettings();
    flushMicrotasks();
    expect(spy).toHaveBeenCalledWith('settings-main');
    spy.calls.mostRecent().returnValue.then((result) => {
      expect(result).toEqual('PUSH');
    });
  }));

});
